Fitur Login dan Register

+++ hapus.php
@@ -1,6 +1,10 @@
 <?php
 include 'koneksi.php'; //menghubungkan dengan database
-
+session_start();
+if (!isset($_SESSION['username'])) {
+    header("Location: login.php");
+    exit;
+}
 $nim = $_GET['id'];
 
 $result = mysqli_query($koneksi, "DELETE from mahasiswa WHERE nim='$nim' ");
